Fix or ignore various Clippy lints

+++ src/signals.rs
@@ -111,7 +111,7 @@ op_future! {
         submission.read_at(fd.fd, ptr, size_of::<libc::signalfd_siginfo>() as u32, NO_OFFSET);
     },
     map_result: |this, (info,), n| {
-        debug_assert_eq!(n, size_of::<libc::signalfd_siginfo>() as i32);
+        debug_assert_eq!(n as usize, size_of::<libc::signalfd_siginfo>());
         // SAFETY: the kernel initialised the info allocation for us as part of
         // the read call.
         Ok(unsafe { Box::<MaybeUninit<libc::signalfd_siginfo>>::assume_init(info) })
